requirements.txt
numpy==1.19.0
ffmpeg_python==0.2.0
dlib==19.20.0
aiogram==2.9.2